Changes made to the DiscoverTheAIContainer component for card rendering and post pagination logic.
In the original code, the RenderCards function was defined within the 'DiscoverTheAIContainer' component.
After the code changed, the RenderCards function was extracted from the main component and defined separately. This helps with modularity and reusability of the card rendering component. In the original code, the logic for rendering the cards in grids (21 columns) and displaying the pagination was more complex and only rendered 21 photos and repeated them.
In the changed code, the pagination logic has been simplified by using the slice method to divide the post array into pages of 21 posts each. The changed code also avoids creating an array of indexes to map over the grids, as the number of grids is calculated directly based on the number of posts.

The dataToUse variable has been introduced, which is used to determine whether search results or all posts should be displayed based on the search state. In terms of performance, it became more efficient, as it avoids creating an array of unnecessary indexes and simplifies the card rendering logic.

Made changes to the 'globalContext.jsx' file a new state for password visibility was created - A new showPassword state variable was introduced using the useState hook to track whether the password should be shown or hidden. Created a function to toggle password visibility togglePasswordVisibility was added to toggle the showPassword value when the user interacts with the password field. The showPassword state and togglePasswordVisibility function have been added to the context values object to make them accessible throughout the application through context.

The useContext import has been removed because it is no longer needed after the changes.

'In the Header.jsx component, the 'Generate' and 'Log In' links have been removed from the non-logged-in user check, as they were not part of the unauthenticated user flow. The 'Log In' link is now displayed directly in the navigation menu for unauthenticated users, without the need to open a login modal first. The openLoginModal and openRegisterModal function calls have been removed from the component, as they are no longer required to open the login and registration modals.'

In the components 'LoginPage.jsx', and 'globalContext.jsx' a new state called showPassword was introduced using useContext, which tracks whether the password should be shown or hidden.

The togglePasswordVisibility function has been added to toggle password visibility when the user interacts with the eye icon. The showPassword state and the togglePasswordVisibility function have been added to the context object to allow the component to use them centrally.

The showPassword state is used to determine whether the password should be displayed as plain text or hidden. The togglePasswordVisibility function is called when the user interacts with the eye icon to toggle password visibility. These changes allow the user to control password visibility in the input field, providing a more interactive and intuitive experience during the login process.

n the ButtonStartCreating.jsx component, the 'MyContext' context and the 'useContext' function were imported to access the global state of the application and the available functions. Use of the 'useContext' hook to obtain context information, such as the user's state and the 'setShowLogin' function to control the display of the login modal. Implementation of the 'handleClick' function to determine the behavior of the button when clicked: If the user is logged in, redirects to the /createpost route. If the user is not logged in, display the login modal setting the showLogin state to true. The <Link> link component has been replaced by a <div> with the ellipse-group class, maintaining the visual appearance of the button. The redirection or display logic of the login modal is handled in the handleClick function.

Changes made to the 'createaccountpage.scss' code, the justify-content property: space-evenly; was added to evenly distribute the space between child elements in the main direction (vertically in this case) of the .createaccountpage class. Added conditional styles for the .user-name1 element when it is valid or has focus. This is done using the :valid and :focus pseudo-selectors. When the field is valid or has focus, the font size and text color are adjusted. Added a style rule for the .toggle-password class, adjusting the icon color.