Changes made to the 'global.scss' file A new font style called ‘Pathway Gothic One’ has been added using @font-face to define the font and its corresponding font file.
Variables have been added to control the duration and delay of animations using the '--s' and '--d' variables.

Changes made to '_variables.scss' , The variables were defined directly within the root :root of the CSS document.

Previously, variables were defined as Sass variables ‘($)’. Now, they have been converted to CSS variables ‘(--)’.

Two new color variables have been added: --color_10 and --background-color_7.

The variable --color_10 was given the value ‘var(--color-lime)’.

The variable --background-color_7 was given the value ‘var(--color-background-button)’.

Added Animation Variables, '--s' and '--d' to control the duration and delay of animations.

In the TextToImageGenerator component 'showImage' and 'showPreview' states were added to control the display of the generated image and the preview.

Changes to the TextToImageGenerator.scss file, replacement of direct style variables with custom CSS variables.

In the original code, the properties for color, font size, font family, and other styles were directly set to values.

In the changed code, these values were replaced with custom CSS variables such as ‘var(--color_1)’, ‘var(--font-family_1)’, etc.

In the Loader component, it replaces the SVG animation with a dot and ring animation using divs and CSS. It creates a series of divs representing animated dots, giving the impression of a load in progress.

In the ‘Loader.scss’ file we introduce a new appearance and behavior for the Loader component.

The original code defined a Loader using SVG and CSS animations to create a rotating circle.

In the new code, the Loader has been completely redefined using divs and CSS to create an animation of dots in a ring.

The new Loader is made up of nested divs to represent points in a ring, creating a different loading animation.
Each point is a div with a rocking animation.

The ring is represented by a div with a solid border.

Variables were defined in the ‘variables.scss’ file and imported to be used in ‘Loader.scss’. This allows for better organization and reuse of colors and other values.

Animations were set to rotate the points and create the loading effect.

The Loader's appearance has been customized with new colors, sizes, and styles.