Creation of loader with App symbol.
Changes made to the LoginPage.jsx file introduce two new Loader variants: a static Loader and an animated Loader, to recreate the application logo using scss and generate a looping animation

'StaticLoader' renders a static Loader with the various rings.

AnimatedLoader renders an animated Loader with the various rings.

The HTML structure has been updated to reflect changes in the visual representation of Loaders.

The new Loaders are made up of multiple divs nested to create the rings and center of the Loader.

CSS classes have been defined in the Loader.scss file to style the new Loaders according to their visual representations.

In AnimatedLoader, several animations '(keyframes)' were defined that will be applied to the loader elements. Each animation is specified with a name and a sequence of transformations that will be applied to the elements.

The .loader class is the main loader container. It defines the size, shape and position of the loader.

Within the .loader class, there are two sub-classes: .static and .Animations.
.static is the static part of the loader, while .Animations is the animated part.

Static part of the '(.static)' loader, multiple static rings are defined with different sizes and styles. They are positioned absolutely relative to the main '(.loader)' container. Each ring has its own class, like .static-ring-7, .static-ring-6, etc. Each ring is drawn using edge and color properties, and they all have a specific 'rotate' defined.

Animated part of the loader '(.Animations)' the same static rings are reset, but now they have animations applied. The animation is defined using the previously defined keyframes. Each ring has its own animation, duration time and rotation direction.

The new Loader components have been separated from the original component to facilitate code reuse and maintenance.

The separation of Loaders into individual components was to facilitate the implementation of specific logic for each type of Loader, in the 'TextToImageGenerator.jsx' component

In 'TextToImageGenerator.jsx', the StaticLoader and AnimatedLoader components are now imported.

The StaticLoader and AnimatedLoader components are used to provide visual loading cues.

StaticLoader is used when the image has not yet been generated and AnimatedLoader is used while the image is being generated.

A negative generatingImg check was created to decide whether StaticLoader should be displayed, while AnimatedLoader is displayed when generatingImg is true.